Handle piece movement in Piece, instead of Board
Converted Board.move_piece_to(piece, pos) to Piece.move_to(pos) for
better ergonomics. A reference to the piece was already necessary, and
the piece itself was already aware of the board it was in. Now you no
longer have to ensure that the piece is actually within the board you
call the movement on.