Allow move when dependent move puts king in danger
Added two more values to the Piece.MovementOutcome enum, a variant of
AVAILABLE and CAPTURE respectively that indicate they put the  current
player's King in danger.

Certain moves, like the pawn two-move advance, are dependent on other
moves - in this instance, the one-move advance. When the King was under
check, the single-move advance might be unavailable because they didn't
put the King into safety. The two-move advance used to see this as
failing the precondition for its own movement, and would also assume
they were impossible.

The additional enum values let us recall that a move may be technically
possible in a vacuum, but unavailable for play because they do not
ensure the safety of the King.