Prevent large movements in far-away scared pieces
Threatened pieces would limit their shuffle away from the attacker with
a clamp on both vector components. This meant that a vertically far-away
piece would shuffle horizontally a ton, even if the relative movement
from the attacker was relatively slight compared to the distance.

This has been replaced with a cap on the movement based on the offset
normal, not the vector component difference